refactor: 로직과 styles 분리
