+++ src/Component/Common/ErrorBox/ErrorBox.styles.ts
@@ -1,0 +1,16 @@
+import styled from "@emotion/styled";
+
+export const StyledDiv = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  height: 40px;
+  background-color: #ffc9c9;
+`;
+
+export const StyledSpan = styled.span`
+  font-weight: 500;
+  font-size: 12px;
+  color: #0a0d13;
+`;
