+++ src/Component/Common/SelectBox/SelectBox.styles.ts
@@ -1,0 +1,9 @@
+import styled from "@emotion/styled";
+
+export const StyledSelect = styled.select`
+  width: 125px;
+  height: 36px;
+  border-radius: 8px;
+  border: 1px solid #0000001a;
+  padding: 8px;
+`;
