--- src/Component/Layout/Body.tsx
@@ -1,20 +1,0 @@
-import { ComponentProps, PropsWithChildren } from "react";
-import styled from "@emotion/styled";
-
-export default function Body({
-  children,
-  ...props
-}: PropsWithChildren<ComponentProps<"div">>) {
-  return <StyledBody {...props}>{children}</StyledBody>;
-}
-
-const StyledBody = styled.div`
-  padding: 36px 24px;
-  margin: 0px;
-  display: flex;
-  flex-direction: column;
-  gap: 24px;
-  height: calc(100vh - 64px);
-  box-sizing: border-box;
-  overflow-y: auto;
-`;
