+++ src/Component/Layout/Body/Body.styles.ts
@@ -1,0 +1,12 @@
+import styled from "@emotion/styled";
+
+export const StyledBody = styled.div`
+  padding: 36px 24px;
+  margin: 0px;
+  display: flex;
+  flex-direction: column;
+  gap: 24px;
+  height: calc(100vh - 64px);
+  box-sizing: border-box;
+  overflow-y: auto;
+`;
