+++ src/Component/Layout/CountBox/CountBox.styles.ts
@@ -1,0 +1,14 @@
+import styled from "@emotion/styled";
+
+export const StyledCountBox = styled.div`
+  width: 19px;
+  height: 19px;
+  background-color: white;
+  border-radius: 50%;
+  position: absolute;
+  left: 7px;
+  bottom: -4px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`;
