+++ src/Component/Layout/CountBox/CountBox.tsx
@@ -1,0 +1,11 @@
+import { StyledCountBox } from "./CountBox.styles";
+
+interface CountBoxProps {
+  children: React.ReactNode;
+}
+
+export default function CountBox({ children }: CountBoxProps) {
+  return (
+    <StyledCountBox data-testid="cart-count-box">{children}</StyledCountBox>
+  );
+}
