+++ src/Component/Layout/Header/Header.styles.ts
@@ -1,0 +1,34 @@
+import styled from "@emotion/styled";
+
+export const StyledHeader = styled.header`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 0 24px;
+  background-color: #000000;
+  height: 64px;
+`;
+
+export const StyledSpan = styled.span`
+  color: #ffffff;
+  font-weight: 800px;
+`;
+
+export const StyledButton = styled.button`
+  width: 24px;
+  height: 24px;
+  padding: 0px;
+  background-color: transparent;
+  border: none;
+  cursor: pointer;
+  position: relative;
+`;
+
+export const StyledImg = styled.img`
+  width: 24px;
+  height: 24px;
+`;
+
+export const StyledCountText = styled.span`
+  font-weight: 700;
+`;
