+++ src/Component/Product/ProductList/ProductList.styles.ts
@@ -1,0 +1,9 @@
+import styled from "@emotion/styled";
+
+export const StyledUl = styled.ul`
+  display: grid;
+  list-style-type: none;
+  padding: 0;
+  grid-template-columns: 1fr 1fr;
+  gap: 16px;
+`;
