+++ src/Component/Product/ProductListContainer/ProductListContainer.styles.ts
@@ -1,0 +1,9 @@
+import styled from "@emotion/styled";
+
+export const StyledSpinnerWrapper = styled.div`
+  width: 100%;
+  height: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+`;
